requirements.txt
fastapi~=0.115.8
uvicorn
sqlalchemy~=2.0.38
pymysql
python-dotenv~=1.0.1
alembic~=1.14.1
email-validator
pydantic[email]~=2.10.6
python-jose[cryptography]
passlib~=1.7.4